# improvement: to_offset avoids binary search if the timestamp is greatr than the biggest one