The subsequent optimization of https://github.com/BurntSushi/jiff/pull/103

## The logic inside of `Tzif#to_offset`

binary_search_by_key find the index from transitions, when the timestamp is greater than all timestamp in transitions. It's not needed to binary search.

It always happens when some timezones had DST before, but DST is canceled for now.

## Change

Compare the timestamp with the biggest timestamp in `transitions`. If `timestamp > self.transitions.last().unwrap().timestamp`, we can avoid the binary search.


## Performance change

In biref, the performance of `instant_to_civil_datetime_static-Asia/Shanghai`  is improved 2 times via cthe urrent PR and #104 

The performance of `instant_to_civil_datetime_static-Asia/Shanghai `:

- `99.3±1.35ns` : Without current PR and #104
- `68.4±5.85ns` : master with #104
- `50.1±1.02ns` : With current PR and #104

```
# before this PR and https://github.com/BurntSushi/jiff/pull/104(master branch)
group                                                base/chrono-tzfile/                    base/chrono/                           base/jiff/
-----                                                -------------------                    ------------                           ----------
instant_to_civil_datetime_static-Asia/Shanghai       1.00     34.7±0.51ns        ? ?/sec    1.08     37.4±0.62ns        ? ?/sec    2.87     99.3±1.35ns        ? ?/sec

# with https://github.com/BurntSushi/jiff/pull/104
group                                                base/chrono-tzfile/                    base/chrono/                           base/jiff/
-----                                                -------------------                    ------------                           ----------
instant_to_civil_datetime_static-Asia/Shanghai       1.00     34.5±0.98ns        ? ?/sec    1.10     38.1±2.72ns        ? ?/sec    1.98     68.4±5.85ns        ? ?/sec


# https://github.com/BurntSushi/jiff/pull/104 with current PR
group                                                base/chrono-tzfile/                    base/chrono/                           base/jiff/
-----                                                -------------------                    ------------                           ----------
instant_to_civil_datetime_static-Asia/Shanghai       1.00     34.6±1.09ns        ? ?/sec    1.15     39.7±6.03ns        ? ?/sec    1.45     50.1±1.02ns        ? ?/sec

```